requirements.txt
django==3.2
djangorestframework==3.13.1
django-debug-toolbar==3.2.1
django-phonenumber-field==6.1.0
phonenumbers==8.12.46
Pillow==8.2.0
environs[django]==9.3.2
requests==2.27.1
geopy==2.2.0
